+++ doc/decision-records/DR003 - queries.md
@@ -14,6 +14,12 @@
 - storage !! => stockage persisté de l'app
 - BookingWrapper => voir ce qu'on fait avec le contexte de booking & useBookingContext
 - useReviewInAppInformation => app state persisté à regarder
+- setFirebaseParams
+- usePersistCookieConsentMutation => c'est une mutation qui appelle le storage
+- useCookies
+- startTrackingAcceptedCookies et server state ?
+- startTracking
+- setMarketingParams => peut etre ?
 
 ## De-hookification
 
@@ -21,3 +27,8 @@
 - useBookingStock => hook de hook
 - useRotatingText => challenger l'animation du loader (on imagine que c'est pour les billeteries externes)
 - reducer dans features/bookOffer, mimique Redux
+- useCookiesChoiceByCategory => ça commence par "use" mais ça n'appelle aucun hook
+
+### Devrait utiliser react-query
+
+- getCookiesLastUpdate => useIsCookiesListUpToDate est plus complexe que nécessaire
