+++ src/features/cookies/helpers/cookiesConsentKey.ts
@@ -1,0 +1,1 @@
+export const COOKIES_CONSENT_KEY = 'cookies'
